refactor: Organize EditHistory module and finish polish pass
Move edit history into Sources/Flux2App/EditHistory/ with split capture/restore
extensions, agent architecture doc, and route-resolver tests. Apply loaded
history pointer on project open, prune status feedback, fill-slider undo on
mouseup, VM smoke history_steps marker, and sidebar file rename.
