+++ Sources/Flux2App/EditHistory/EditHistorySidebarSection.swift
